+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added `approval_required` and `invites_enabled` flags to `/api/v1/instance` endpoint response.
+
 ### Removed
 
 - Dropped support for `blockchain` configuration parameter.