+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Dropped support for `blockchain` configuration parameter.
 
+### Fixed
+
+- Added missing `<link rel="self">` element to Atom feeds.
+
 ## [1.11.0] - 2023-01-23
 
 ### Added