+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `approval_required` and `invites_enabled` flags to `/api/v1/instance` endpoint response.
 - Added `registration.type` configuration option (replaces `registrations_open`).
+- Implemented roles & permissions.
+- Added "read-only user" role.
 
 ### Deprecated