+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.12.0] - 2023-01-26
+
 ### Added
 
 - Added `approval_required` and `invites_enabled` flags to `/api/v1/instance` endpoint response.